requirements.txt
firebase-admin
pyrebase
customtkinter
Pillow
google-auth-oauthlib
google-auth
requests
tkinterdnd2